refactor: styling in sidebar
