+++ apps/service-portal/src/components/Sidebar/NavItem/NavItem.css.ts
@@ -1,21 +1,38 @@
-import { style, styleVariants, keyframes } from '@vanilla-extract/css'
+import { style, styleVariants, globalStyle } from '@vanilla-extract/css'
 import { theme, themeUtils } from '@island.is/island-ui/theme'
-
-const iconEaseIn = keyframes({
-  from: {
-    opacity: 0.8,
-  },
-  to: {
-    opacity: 1,
-  },
-})
 
 export const dotState = styleVariants({
   active: {},
   inactive: {},
 })
 
-export const navItem = style({})
+export const navItem = style({
+  ':hover': {
+    color: theme.color.blue400,
+    backgroundColor: theme.color.blue100,
+  },
+})
+
+globalStyle(`${navItem}:hover svg`, {
+  color: theme.color.blue400,
+})
+
+globalStyle(`${navItem}:hover #sub-nav-model`, {
+  display: 'flex',
+  position: 'absolute',
+  top: 0,
+  left: 28,
+  width: 220,
+  height: 'max-content',
+  zIndex: 10,
+})
+
+export const inner = style({
+  left: 64,
+  width: 183,
+  borderRadius: theme.border.radius.standard,
+  transition: 'all 250ms ease-in-out',
+})
 
 export const navItemActive = styleVariants({
   active: {
@@ -61,28 +78,6 @@ export const navItemActive = styleVariants({
   },
 })
 
-export const navItemHover = styleVariants({
-  hoverActive: {
-    color: theme.color.blue400,
-    borderLeft: `4px solid ${theme.color.blue400}`,
-    textDecoration: 'none',
-  },
-  hoverInactive: {
-    backgroundColor: theme.color.blue100,
-    color: theme.color.blue400,
-    borderLeft: `4px solid ${theme.color.blue100}`,
-  },
-  hoverCollapsed: {
-    ...themeUtils.responsiveStyle({
-      lg: {
-        backgroundColor: theme.color.blue100,
-        color: theme.color.blue400,
-        borderRadius: '8px',
-      },
-    }),
-  },
-  none: {},
-})
 export const text = style({
   fontSize: 16,
   lineHeight: '26px',
