+++ apps/service-portal/src/components/Sidebar/NavItem/NavItem.tsx
@@ -7,20 +7,13 @@ import { useStore } from '../../../store/stateProvider'
 import { useWindowSize } from 'react-use'
 import { theme } from '@island.is/island-ui/theme'
 import cn from 'classnames'
-
-interface ChevronProps {
-  expanded?: boolean
-  enabled?: boolean
-  active: boolean
-  onChevronClick?: () => void
-}
+import Chevron from './Chevron'
 
 interface Props {
   path?: ServicePortalPath
   icon?: Pick<IconProps, 'icon' | 'type'>
   active: boolean
   chevron?: boolean
-  hover: boolean
   enabled?: boolean
   expanded?: boolean
   external?: boolean
@@ -34,7 +27,6 @@ interface Props {
 const NavItemContent: FC<Props> = ({
   icon,
   active,
-  hover,
   enabled,
   hasArray,
   onClick,
@@ -55,25 +47,11 @@ const NavItemContent: FC<Props> = ({
     ? 'inactiveCollapsed'
     : 'inactive'
 
-  let navItemHover: keyof typeof styles.navItemHover = 'none'
-
-  if (hover && active) {
-    navItemHover = 'hoverActive'
-  } else if (hover && collapsed) {
-    navItemHover = 'hoverCollapsed'
-  } else if (hover && !active) {
-    navItemHover = 'hoverInactive'
-  } else {
-    navItemHover = 'none'
-  }
   const badgeActive: keyof typeof styles.badge = badge ? 'active' : 'inactive'
 
   return (
     <Box
-      className={[
-        styles.navItemHover[navItemHover],
-        styles.navItemActive[navItemActive],
-      ]}
+      className={[styles.navItem, styles.navItemActive[navItemActive]]}
       display="flex"
       alignItems="center"
       justifyContent={collapsed ? 'center' : 'spaceBetween'}
@@ -108,7 +86,7 @@ const NavItemContent: FC<Props> = ({
             <Icon
               type={active ? 'filled' : 'outline'}
               icon={icon.icon}
-              color={active || hover ? 'blue400' : 'blue600'}
+              color={active ? 'blue400' : 'blue600'}
               size="medium"
               className={styles.icon}
             />
@@ -121,40 +99,12 @@ const NavItemContent: FC<Props> = ({
           type="filled"
           icon="lockClosed"
           size="small"
-          color={hover || active ? 'blue400' : 'blue600'}
+          color={active ? 'blue400' : 'blue600'}
           className={cn(styles.lock, collapsed && styles.lockCollapsed)}
         />
       )}
     </Box>
   )
-}
-
-const Chevron: FC<ChevronProps> = ({
-  expanded,
-  enabled,
-  active,
-  onChevronClick,
-}) => {
-  const chevronIcon = expanded ? 'chevronUp' : 'chevronDown'
-  const [{ sidebarState }] = useStore()
-  const showChevron = enabled && sidebarState === 'open'
-
-  return showChevron ? (
-    <Box
-      onClick={() => {
-        onChevronClick && onChevronClick()
-      }}
-      className={styles.iconWrapper}
-    >
-      <Icon
-        type="filled"
-        icon={chevronIcon}
-        size="medium"
-        color={active ? 'blue400' : 'blue600'}
-        className={styles.icon}
-      />
-    </Box>
-  ) : null
 }
 
 const NavItem: FC<Props> = (props) => {
